project/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
asyncpg==0.27.0
sqlalchemy==2.0.15
alembic==1.11.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.7
python-multipart==0.0.6
PyJWT==2.10.1